move System.Net.ClientModel.Tests in
