+++ sdk/core/System.Net.ClientModel/tests/ModelReaderWriter/Models/AvailabilitySetDataTests.cs
@@ -4,21 +4,16 @@
 using NUnit.Framework;
 using System.Collections.Generic;
 using System.IO;
-using System.Net.ClientModel.Core;
 using System.Net.ClientModel.Tests.Client;
-using System.Net.ClientModel.Tests.Client.ResourceManager.Compute;
+using System.Net.ClientModel.Tests.Client.Models.ResourceManager.Compute;
 
-namespace System.Net.ClientModel.Tests.ModelReaderWriterTests
+namespace System.Net.ClientModel.Tests.ModelReaderWriterTests.Models
 {
-    internal class AvailabilitySetDataTestsWithVMs : ModelJsonTests<AvailabilitySetData>
+    internal class AvailabilitySetDataTests : ModelJsonTests<AvailabilitySetData>
     {
-        protected override string WirePayload => File.ReadAllText(TestData.GetLocation("AvailabilitySetData/AvailabilitySetDataWithVMsWireFormat.json")).TrimEnd();
+        protected override string WirePayload => File.ReadAllText(TestData.GetLocation("AvailabilitySetData/AvailabilitySetDataWireFormat.json")).TrimEnd();
 
         protected override string JsonPayload => WirePayload;
-
-        protected override Func<AvailabilitySetData?, MessageBody> ToPipelineContent => model => model;
-
-        protected override Func<Result?, AvailabilitySetData> FromResult => response => (AvailabilitySetData)response;
 
         protected override string GetExpectedResult(ModelReaderWriterFormat format)
         {
@@ -28,11 +23,10 @@ namespace System.Net.ClientModel.Tests.ModelReaderWriterTests
             expectedSerializedString += "\"sku\":{\"name\":\"Classic\"";
             //if (!ignoreAdditionalProperties)
             //    expectedSerializedString += ",\"extraSku\":\"extraSku\"";
-            expectedSerializedString += "},\"tags\":{\"key\":\"value\"},\"location\":\"eastus\",\"properties\":{\"platformUpdateDomainCount\":5,\"platformFaultDomainCount\":3";
+            expectedSerializedString += "},\"tags\":{\"key\":\"value\"},\"location\":\"eastus\",\"properties\":{\"platformUpdateDomainCount\":5,\"platformFaultDomainCount\":3}";
             //if (!ignoreAdditionalProperties)
             //    expectedSerializedString += ",\"extraRoot\":\"extraRoot\"";
-            expectedSerializedString += ",\"virtualMachines\":[{\"id\":\"/subscriptions/e37510d7-33b6-4676-886f-ee75bcc01871/resourceGroups/testRG-6497/providers/Microsoft.Compute/availabilitySets/testAS1\"},{\"id\":\"/subscriptions/e37510d7-33b6-4676-886f-ee75bcc01871/resourceGroups/testRG-6497/providers/Microsoft.Compute/availabilitySets/testAS2\"}]";
-            expectedSerializedString += "}}";
+            expectedSerializedString += "}";
             return expectedSerializedString; ;
         }
 
@@ -48,9 +42,6 @@ namespace System.Net.ClientModel.Tests.ModelReaderWriterTests
             Assert.AreEqual(5, model.PlatformUpdateDomainCount);
             Assert.AreEqual(3, model.PlatformFaultDomainCount);
             Assert.AreEqual("Classic", model.Sku.Name);
-            Assert.AreEqual(2, model.VirtualMachines.Count);
-            Assert.AreEqual("/subscriptions/e37510d7-33b6-4676-886f-ee75bcc01871/resourceGroups/testRG-6497/providers/Microsoft.Compute/availabilitySets/testAS1", model.VirtualMachines[0].Id.ToString());
-            Assert.AreEqual("/subscriptions/e37510d7-33b6-4676-886f-ee75bcc01871/resourceGroups/testRG-6497/providers/Microsoft.Compute/availabilitySets/testAS2", model.VirtualMachines[1].Id.ToString());
         }
 
         protected override void CompareModels(AvailabilitySetData model, AvailabilitySetData model2, ModelReaderWriterFormat format)
@@ -64,9 +55,6 @@ namespace System.Net.ClientModel.Tests.ModelReaderWriterTests
                 Assert.AreEqual(model.ResourceType, model2.ResourceType);
             CollectionAssert.AreEquivalent(model.Tags, model2.Tags);
             Assert.AreEqual(model.Sku.Name, model2.Sku.Name);
-            Assert.AreEqual(model.VirtualMachines.Count, model2.VirtualMachines.Count);
-            Assert.AreEqual(model.VirtualMachines[0].Id, model2.VirtualMachines[0].Id);
-            Assert.AreEqual(model.VirtualMachines[1].Id, model2.VirtualMachines[1].Id);
         }
     }
 }
