+++ sdk/core/System.Net.ClientModel/tests/ModelReaderWriter/Models/ResourceProviderDataTests.cs
@@ -3,21 +3,16 @@
 
 using NUnit.Framework;
 using System.IO;
-using System.Net.ClientModel.Core;
 using System.Net.ClientModel.Tests.Client;
-using System.Net.ClientModel.Tests.Client.ResourceManager.Resources;
+using System.Net.ClientModel.Tests.Client.Models.ResourceManager.Resources;
 
-namespace System.Net.ClientModel.Tests.ModelReaderWriterTests
+namespace System.Net.ClientModel.Tests.ModelReaderWriterTests.Models
 {
     internal class ResourceProviderDataTests : ModelJsonTests<ResourceProviderData>
     {
         protected override string JsonPayload => WirePayload;
 
         protected override string WirePayload => File.ReadAllText(TestData.GetLocation("ResourceProviderData/ResourceProviderData-Collapsed.json")).TrimEnd();
-
-        protected override Func<ResourceProviderData?, MessageBody> ToPipelineContent => model => model;
-
-        protected override Func<Result?, ResourceProviderData> FromResult => response => (ResourceProviderData)response;
 
         protected override void CompareModels(ResourceProviderData model, ResourceProviderData model2, ModelReaderWriterFormat format)
         {
