+++ sdk/core/System.Net.ClientModel/tests/ModelReaderWriter/Models/UnknownBaseModelTests.cs
@@ -3,10 +3,9 @@
 
 using NUnit.Framework;
 using System.Linq;
-using System.Net.ClientModel.Core;
 using System.Net.ClientModel.Tests.Client.ModelReaderWriterTests.Models;
 
-namespace System.Net.ClientModel.Tests.ModelReaderWriterTests
+namespace System.Net.ClientModel.Tests.ModelReaderWriterTests.Models
 {
     internal class UnknownBaseModelTests : ModelJsonTests<BaseModel>
     {
@@ -19,10 +18,6 @@ namespace System.Net.ClientModel.Tests.ModelReaderWriterTests
         protected override string JsonPayload => WirePayload;
 
         protected override string WirePayload => "{\"kind\":\"Z\",\"name\":\"zmodel\",\"zProperty\":1.5,\"extra\":\"stuff\"}";
-
-        protected override Func<BaseModel?, MessageBody> ToPipelineContent => model => model;
-
-        protected override Func<Result?, BaseModel> FromResult => result => (BaseModel)result;
 
         protected override void CompareModels(BaseModel model, BaseModel model2, ModelReaderWriterFormat format)
         {
